+++ saleor/extensions/manager.py
@@ -252,12 +252,6 @@ class ExtensionsManager(PaymentInterface):
         method_name = "process_payment"
         return self.__run_payment_method(gateway, method_name, payment_information)
 
-    def create_payment_form(self, data, gateway, payment_information):
-        method_name = "create_form"
-        return self.__run_payment_method(
-            gateway, method_name, payment_information, data=data
-        )
-
     def get_client_token(self, gateway, token_config: "TokenConfig") -> str:
         method_name = "get_client_token"
         default_value = None