+++ saleor/payment/gateway.py
@@ -191,13 +191,6 @@ def confirm(payment: Payment) -> Transaction:
     )
 
 
-@require_active_payment
-def create_payment_form(payment: Payment, data) -> forms.Form:
-    plugin_manager = get_extensions_manager()
-    payment_data = create_payment_information(payment)
-    return plugin_manager.create_payment_form(data, payment.gateway, payment_data)
-
-
 def list_payment_sources(gateway: str, customer_id: str) -> List["CustomerSource"]:
     plugin_manager = get_extensions_manager()
     return plugin_manager.list_payment_sources(gateway, customer_id)