+++ saleor/payment/gateways/dummy/plugin.py
@@ -10,7 +10,6 @@ from . import (
     authorize,
     capture,
     confirm,
-    create_form,
     get_client_token,
     process_payment,
     refund,
@@ -21,7 +20,6 @@ GATEWAY_NAME = "Dummy"
 
 if TYPE_CHECKING:
     from ...interface import GatewayResponse, PaymentData, TokenConfig
-    from django import forms
 
 
 def require_active_plugin(fn):
@@ -38,13 +36,6 @@ def require_active_plugin(fn):
 class DummyGatewayPlugin(BasePlugin):
     PLUGIN_NAME = GATEWAY_NAME
     CONFIG_STRUCTURE = {
-        "Template path": {
-            "type": ConfigurationTypeField.STRING,
-            "help_text": pgettext_lazy(
-                "Plugin help text", "Location of django payment template for gateway."
-            ),
-            "label": pgettext_lazy("Plugin label", "Template path"),
-        },
         "Store customers card": {
             "type": ConfigurationTypeField.BOOLEAN,
             "help_text": pgettext_lazy(
@@ -69,7 +60,6 @@ class DummyGatewayPlugin(BasePlugin):
             gateway_name=GATEWAY_NAME,
             auto_capture=True,
             connection_params={},
-            template_path="",
             store_customer=False,
         )
 
@@ -84,7 +74,6 @@ class DummyGatewayPlugin(BasePlugin):
                 gateway_name=GATEWAY_NAME,
                 auto_capture=configuration["Automatic payment capture"],
                 connection_params={},
-                template_path=configuration["Template path"],
                 store_customer=configuration["Store customers card"],
             )
 
@@ -97,7 +86,6 @@ class DummyGatewayPlugin(BasePlugin):
             "configuration": [
                 {"name": "Store customers card", "value": False},
                 {"name": "Automatic payment capture", "value": True},
-                {"name": "Template path", "value": "order/payment/dummy.html"},
             ],
         }
         return defaults
@@ -140,12 +128,6 @@ class DummyGatewayPlugin(BasePlugin):
         self, payment_information: "PaymentData", previous_value
     ) -> "GatewayResponse":
         return process_payment(payment_information, self._get_gateway_config())
-
-    @require_active_plugin
-    def create_form(
-        self, data, payment_information: "PaymentData", previous_value
-    ) -> "forms.Form":
-        return create_form(data, payment_information, {})
 
     @require_active_plugin
     def get_client_token(self, token_config: "TokenConfig", previous_value):